General warning cleanup 1
* A lot of hit has to do with size_t being unsigned long long in 64 bit.
  - Just explicitly casting for a lot of cases removes the warning
* Good bit are also the differences in struct definitions to match packets and function definitions
  - Also just cast to fix
* Used a lot of #pragma warning( disable : 4244/4267 ) for template warnings
* InviteHandlers.cpp line 118 was definitely a typo bug. Needed assignment "=" instead of "=="

+++ deps/datReader/Dat.cpp
@@ -292,7 +292,7 @@ namespace xiv::dat
     DatBlockHeader block_header = extract< DatBlockHeader >( m_handle );
 
     // Resizing the vector to write directly into it
-    const uint32_t data_size = o_data.size();
+    const uint32_t data_size = (uint32_t)o_data.size();
     o_data.resize( data_size + block_header.uncompressed_size );
 
     // 32000 in compressed_size means it is not compressed so take uncompressed_size
@@ -308,7 +308,7 @@ namespace xiv::dat
       m_handle.read( temp_buffer.data(), block_header.compressed_size );
 
       utils::zlib::no_header_decompress( reinterpret_cast<uint8_t*>(temp_buffer.data()),
-                                         temp_buffer.size(),
+                                         (uint32_t)temp_buffer.size(),
                                          reinterpret_cast<uint8_t*>(o_data.data() + data_size),
                                          block_header.uncompressed_size );
     }
