+++ deps/mysqlConnector/Statement.cpp
@@ -17,7 +17,7 @@ Mysql::Statement::Statement( std::shared_ptr< Mysql::Connection > conn ) :
 
 void Mysql::Statement::doQuery( const std::string &q )
 {
-   mysql_real_query( m_pConnection->getRawCon(), q.c_str(), q.length() );
+   mysql_real_query( m_pConnection->getRawCon(), q.c_str(), (unsigned long)q.length() );
 
    if( errNo() )
       throw std::runtime_error( m_pConnection->getError() );
